requirements.txt
fastapi==0.99.1
uvicorn==0.22.0
sqlalchemy==2.0.17
pydantic==1.10.8
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
pillow==9.5.0
psycopg2-binary==2.9.6
pytest==7.3.1
httpx==0.24.1
asyncio==3.4.3
aiofiles==23.1.0
bcrypt==4.0.1
python-jose==3.3.0
passlib==1.7.4
